feat: generate the TypeScript IPC contract from the Rust definitions
Every payload type crossing the Tauri boundary was transcribed by hand into
`src/types/ipc.ts`, which is only ever wrong in one direction: the frontend
compiles against a shape the backend no longer sends. Replace the transcription
with generation, and add a fourth verification gate so the committed artifact
cannot drift.

- Add `src-tauri/src/bindings.rs`, the crate's single command declaration: one
  `tauri_specta::Builder` that both supplies the invoke handler and generates
  `src/types/bindings.ts`. `Builder<R>` is generic, so the same function serves
  `Wry` in `lib.rs` and `MockRuntime` in `ipc_tests.rs` — two lists that merely
  happen to agree is the bug this prevents
- Delete `handlers.rs`; a guard test rejects any `generate_handler!` and any
  second `collect_commands!` list, counting the macro rather than the builder
  type so a concretely-typed `Builder::<Wry>::new()` cannot slip past
- Commit the generated `src/types/bindings.ts` and reduce `src/types/ipc.ts` to
  re-exports plus `isErrorDto`, the one runtime narrowing a generator cannot
  produce, now defined against the generated type
- Add `bindings:check` to `npm run verify` and to CI, with
  `git ls-files --error-unmatch` so an untracked file cannot make the drift gate
  vacuous
- Add a bypass lint rejecting `invoke("name")` outside the generated module;
  it caught a raw invocation in `App.test.tsx`

Two specta constraints now bind every DTO. 64-bit integers cannot cross the
boundary — `latency_ms` becomes `u32`, correcting a hand-written declaration
that claimed a `u64` fits losslessly in a JavaScript number.
`skip_serializing_if` is unavailable under unified serde, so an absent optional
travels as an explicit `null` and is typed `T | null`; the declared type and the
wire now agree exactly, which is the point.

Commands keep rejecting (`ErrorHandlingMode::Throw`) rather than resolving a
tagged result, because the generated `Result` runtime still rethrows
`instanceof Error` and would give every call site two error channels instead of
one. `.typ::<ErrorDto>()` forces the error type into the output, since `Throw`
puts it in no signature.

Archive the change and sync `openspec/specs/typed-ipc/`.

+++ src-tauri/src/commands/config.rs
@@ -28,16 +28,19 @@ const AI_API_KEY: &str = "ai.api_key";
 const CONNECTION_TEST_HINT: &str = "config.connection_test.hint";
 
 #[tauri::command]
+#[specta::specta]
 pub fn get_config(state: State<AppState>) -> Result<ConfigDto, ErrorDto> {
     read_config(state.config_service())
 }
 
 #[tauri::command]
+#[specta::specta]
 pub fn set_config_value(request: SetConfigRequest, state: State<AppState>) -> Result<(), ErrorDto> {
     write_config_value(state.config_service(), &request.key, &request.value)
 }
 
 #[tauri::command]
+#[specta::specta]
 pub async fn test_ai_connection(
     state: State<'_, AppState>,
 ) -> Result<ConnectionTestResult, ErrorDto> {
@@ -138,7 +141,7 @@ async fn probe_ai_connection(service: &dyn ConfigService) -> ConnectionTestResul
     match provider.chat_completion(probe).await {
         Ok(_) => ConnectionTestResult {
             ok: true,
-            latency_ms: Some(started.elapsed().as_millis() as u64),
+            latency_ms: Some(started.elapsed().as_millis() as u32),
             error: None,
         },
         Err(err) => failed_test(err),
